+++ src/builtins/fish_indent.rs
@@ -7,7 +7,6 @@ use crate::{
     common::{PROGRAM_NAME, get_program_name, unescape_string},
     env::{EnvStack, env_init, environment::Environment as _},
     err_fmt, err_str,
-    expand::INTERNAL_SEPARATOR,
     global_safety::RelaxedAtomicBool,
     highlight::{HighlightRole, HighlightSpec, colorize, highlight_shell},
     locale::set_libc_locales,
@@ -26,7 +25,7 @@ use assert_matches::assert_matches;
 use fish_common::{ReadExt as _, UnescapeFlags, UnescapeStringStyle};
 use fish_wcstringutil::count_preceding_backslashes;
 use fish_wgetopt::{ArgType, WGetopter, WOption, wopt};
-use fish_widestring::{bytes2wcstring, osstr2wcstring, wcs2bytes};
+use fish_widestring::{INTERNAL_SEPARATOR, bytes2wcstring, osstr2wcstring, wcs2bytes};
 use std::{
     ffi::OsStr,
     fmt::Write as _,
